Weatherdata/requirements.txt
Django==4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
numpy==1.20.3
openmeteo-requests==1.1.0
pandas==1.2.4
requests-cache==1.1.1
retry-requests==2.0.0
